# docs/architecture/api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Daily Stock Analysis API",
    "description": "A股自選股智能分析系統 API\n\n## 功能模組\n- 股票分析：觸發 AI 智能分析\n- 歷史記錄：查詢歷史分析報告\n- 股票資料：取得行情資料\n\n## 認證方式\n目前版本暫無認證要求",
    "version": "1.0.0",
    "contact": {
      "name": "Daily Stock Analysis Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地開發伺服器"
    }
  ],
...
  "paths": {
...
    "/api/v1/analysis/analyze": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "觸發股票分析",
        "description": "啟動 AI 智能分析任務，支援單支或多支股票批次分析",
        "operationId": "triggerAnalysis",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "分析完成（同步模式）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            }
          },
          "202": {
            "description": "分析任務已接受（非同步模式）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAccepted"
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "股票正在分析中，拒絕重複提交",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateTaskError"
                }
              }
            }
          },
          "500": {
            "description": "分析失敗",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/tasks": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "取得分析任務列表",
        "description": "取得目前所有分析任務，支援依狀態篩選。回傳進行中和最近完成的任務。",
        "operationId": "getAnalysisTasks",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "篩選狀態：pending, processing, completed, failed（支援逗號分隔多個）",
            "schema": {
              "type": "string",
              "example": "pending,processing"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "回傳數量限制",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "任務列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyzeRequest": {
        "type": "object",
        "properties": {
          "stock_code": {
            "type": "string",
            "description": "單支股票代碼",
            "example": "AAPL"
          },
          "stock_codes": {
            "type": "array",
            "description": "多支股票代碼（與 stock_code 二選一）",
            "items": {
              "type": "string"
            },
            "example": [
              "AAPL",
              "TSLA"
            ]
          },
          "report_type": {
            "type": "string",
            "enum": [
              "simple",
              "detailed"
            ],
            "default": "detailed",
            "description": "報告類型"
          },
          "force_refresh": {
            "type": "boolean",
            "default": false,
            "description": "是否強制刷新（忽略快取）"
          },
          "async_mode": {
            "type": "boolean",
            "default": false,
            "description": "是否使用非同步模式"
          }
        }
      },
      "AnalysisResult": {
        "type": "object",
        "properties": {
          "query_id": {
            "type": "string",
            "description": "分析記錄唯一識別碼"
          },
          "stock_code": {
            "type": "string"
          },
          "stock_name": {
            "type": "string"
          },
          "report": {
            "$ref": "#/components/schemas/AnalysisReport"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "query_id",
          "stock_code",
          "report",
          "created_at"
        ]
      },
      "TaskAccepted": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "string",
            "description": "任務 ID，用於查詢狀態"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing"
            ],
            "example": "pending"
          },
          "message": {
            "type": "string",
            "example": "Analysis task accepted"
          }
        },
        "required": [
          "task_id",
          "status"
        ]
      },
...
      "TaskInfo": {
        "type": "object",
        "description": "任務詳情（用於任務列表和 SSE 事件）",
        "properties": {
          "task_id": {
            "type": "string",
            "description": "任務 ID"
          },
          "stock_code": {
            "type": "string",
            "description": "股票代碼"
          },
          "stock_name": {
            "type": "string",
            "description": "股票名稱"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed"
            ],
            "description": "任務狀態"
          },
          "progress": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "description": "進度百分比"
          },
          "message": {
            "type": "string",
            "description": "狀態訊息"
          },
          "report_type": {
            "type": "string",
            "enum": [
              "simple",
              "detailed"
            ],
            "description": "報告類型"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "建立時間"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "description": "開始執行時間"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "description": "完成時間"
          },
          "error": {
            "type": "string",
            "description": "錯誤訊息"
          }
        },
        "required": [
          "task_id",
          "stock_code",
          "status",
          "created_at"
        ]
      },
      "TaskListResponse": {
        "type": "object",
        "description": "任務列表回應",
        "properties": {
          "total": {
            "type": "integer",
            "description": "任務總數"
          },
          "pending": {
            "type": "integer",
            "description": "等待中的任務數"
          },
          "processing": {
            "type": "integer",
            "description": "處理中的任務數"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskInfo"
            },
            "description": "任務列表"
          }
        },
        "required": [
          "total",
          "pending",
          "processing",
          "tasks"
        ]
      },
      "DuplicateTaskError": {
        "type": "object",
        "description": "重複任務錯誤回應",
        "properties": {
          "error": {
            "type": "string",
            "example": "duplicate_task",
            "description": "錯誤類型"
          },
          "message": {
            "type": "string",
            "example": "股票 AAPL 正在分析中",
            "description": "錯誤訊息"
          },
          "stock_code": {
            "type": "string",
            "example": "AAPL",
            "description": "股票代碼"
          },
          "existing_task_id": {
            "type": "string",
            "example": "abc123def456",
            "description": "已存在的任務 ID"
          }
        },
        "required": [
          "error",
          "message",
          "stock_code",
          "existing_task_id"
        ]
      },
...
      "AnalysisReport": {
        "type": "object",
        "description": "完整分析報告",
        "properties": {
          "meta": {
            "type": "object",
            "description": "元資訊",
            "properties": {
              "query_id": {
                "type": "string"
              },
              "stock_code": {
                "type": "string"
              },
              "stock_name": {
                "type": "string"
              },
              "report_type": {
                "type": "string"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "model_used": {
                "type": "string",
                "description": "分析使用的 LLM 模型（完整名稱，如 gemini/gemini-2.0-flash）"
              }
            }
          },
          "summary": {
            "type": "object",
            "description": "概覽區（首屏展示）",
            "properties": {
              "analysis_summary": {
                "type": "string",
                "description": "關鍵結論"
              },
              "operation_advice": {
                "type": "string",
                "description": "操作建議"
              },
              "trend_prediction": {
                "type": "string",
                "description": "趨勢預測"
              },
              "sentiment_score": {
                "type": "integer",
                "description": "情緒評分 (0-100)"
              },
              "sentiment_label": {
                "type": "string",
                "description": "情緒標籤",
                "enum": [
                  "極度悲觀",
                  "悲觀",
                  "中性",
                  "樂觀",
                  "極度樂觀"
                ]
              }
            }
          },
          "strategy": {
            "type": "object",
            "description": "策略點位區",
            "properties": {
              "ideal_buy": {
                "type": "string",
                "description": "理想買入價"
              },
              "secondary_buy": {
                "type": "string",
                "description": "第二買入價"
              },
              "stop_loss": {
                "type": "string",
                "description": "止損價"
              },
              "take_profit": {
                "type": "string",
                "description": "止盈價"
              }
            }
          },
          "details": {
            "type": "object",
            "description": "詳情區（可折疊）",
            "properties": {
              "news_content": {
                "type": "string",
                "description": "新聞摘要"
              },
              "raw_result": {
                "type": "object",
                "description": "原始分析結果（JSON）"
              },
              "context_snapshot": {
                "type": "object",
                "description": "分析時上下文快照（JSON）"
              }
            }
          }
        },
        "required": [
          "meta",
          "summary"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "錯誤類型"
          },
          "message": {
            "type": "string",
            "description": "錯誤詳情"
          },
          "detail": {
            "type": "object",
            "description": "附加錯誤資訊"
          }
        },
        "required": [
          "error",
          "message"
        ]
      }
    }
  }
}